requirements.txt
git+https://github.com/openai/whisper.git
pyannote.audio
torch
torchaudio
torchvision
pydub
ffmpeg-python
